requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
fastapi==0.127.0
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.42.6
locust-cloud==1.30.0
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
platformdirs==4.5.1
pluggy==1.6.0
psutil==7.2.0
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4
pytest-benchmark==5.2.3
python-dotenv==1.2.1
python-engineio==4.13.0
python-socketio==5.16.0
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.4
simple-websocket==1.1.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
python-jose[cryptography]==3.3.0
pwdlib[bcrypt]==0.2.1
pwdlib[argon2]
email-validator==2.2.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
aiosqlite==0.21.0
pytest-asyncio==0.25.3
